chore: refactor email services
